backend/data/api/views.py
from ..models import Document
from .serializers import DocumentSerializer
from rest_framework import generics, permissions

class DocumentListCreate(generics.ListCreateAPIView):
    permission_classes = [permissions.IsAuthenticated]
    serializer_class = DocumentSerializer

    def get_queryset(self):
        return self.request.user.todos.all()

    def perform_create(self, serializer):
        serializer.save(owner=self.request.user)